refactored components and removed HeaderLayout and dropdown
