+++ frontend/src/components/ui/index.ts
@@ -1,1 +1,4 @@
 export { ChatBallon } from "./ChatBalloon";
+export { Menu } from "./Menu";
+export { Input } from "./Input";
+export { Card } from "./Card";
